+++ plugins/gs/config.d.ts
@@ -1,0 +1,10 @@
+export interface Config {
+  /** @visibility frontend */
+  gs: {
+    /** @visibility frontend */
+    endpoints: {
+      /** @visibility frontend */
+      snail: string;
+    }
+  }
+}