+++ plugins/gs/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './GSApi';
+export * from './GSClient';