+++ plugins/gs/src/components/ClustersTable/index.ts
@@ -1,0 +1,1 @@
+export { ClustersTable } from './ClustersTable';